Created CitiesContext component and and enclose app.js children
